requirements.txt
playwright==1.50.0
python-dotenv==1.0.1
pytest==7.3.1
pytest-playwright==0.3.0
